+++ src/index.js
@@ -8,6 +8,25 @@ export { renderBlock, renderInline, renderPlain } from './i18n/markdown.js'
 export { languageLabels, offeredLanguages } from './languages.js'
 export { defaultScrollBehavior } from './router/index.js'
 export { mediaUrl, useSiteConfig } from './siteConfig.js'
+
+// The list pages: query state, pagination, facets, dates, keyword search.
+export { useListQuery } from './catalogue/listQuery.js'
+export { paginate, sortChronological, usePagination } from './catalogue/pagination.js'
+export { facetOptions, useFacets } from './catalogue/facets.js'
+export {
+  dateRange, eraLabel, inDateRange, roundOutward, yearBuckets, yearBucketsFromRange,
+} from './catalogue/dates.js'
+export { parseBooleanQuery, useKeywordIndex } from './catalogue/keywordIndex.js'
+
+// The record page: the sheet engine and its derivations.
+export { glossaryEntries, glossaryTermsFor, sheetRows, useRecordSheet } from './record/recordSheet.js'
+export {
+  citation, relatedRecords, searchGlossary, timelineLinkFor, useGlossaryPopup, useRelatedRecords,
+} from './record/derivations.js'
+
+// Conventions every website had written for itself.
+export { PROJECT_ENTRIES, projectName, useFeaturedRecord, useProjectName, useSection } from './conventions.js'
+
 export { default as I18nText } from './i18n/I18nText.vue'
 export { default as I18nTextInline } from './i18n/I18nTextInline.vue'
 export { default as HomeView } from './views/HomeView.vue'